removing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@4934 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/xml/GenericSAXHandler.cpp
@@ -218,12 +218,12 @@ GenericSAXHandler::endElement(const XMLCh* const /*uri*/,
     SumoXMLTag element = convertTag(name);
     // collect characters
     size_t len = 0;
-    size_t i;
+    unsigned i;
     for (i=0; i<myCharactersVector.size(); ++i) {
         len += myCharactersVector[i].length();
     }
     char *buf = new char[len+1];
-    int pos = 0;
+    size_t pos = 0;
     for (i=0; i<myCharactersVector.size(); ++i) {
         memcpy((unsigned char*) buf+pos, (unsigned char*) myCharactersVector[i].c_str(),
                sizeof(char)*myCharactersVector[i].length());
